Lazy register of SpringConfiguration
In case you use the AnnotatedSpringConfiguration
and construct beans that need the Dropwizard
Configuration or Environment.

Those beans cannot be instantiated when you create
the context, as the registration of the dropwizard
configuration and environment as singletons into the
context happens after the beans would be constructed.

There is now an additional constructor in the SpringBundle
which allows you to hand over the annotated Spring configuration
class, which then gets registered with the context after
those singletons got registered.